Get the WCOFS S3 output for the pervious month
Code to get the previous month's output from WCOFS from the S3 bucket.  This is to create a intermediate pickle file of the data.  This is to help speed up the load of WCOFS data as internet connections can be slow.  These files will be used at the end of the year to create a new NetCDF file to aid in loading the last years data.